# payment-service/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  version: 2.0.0
  description: |
    API for creating and managing payments through the parallel saga state machine.
    Payments flow through FX conversion followed by parallel authorization, fraud, limits,
    sanctions, and fee checks before reaching a terminal state.

paths:
  /api/payments:
    post:
      tags:
        - Payments
      summary: Create a new payment
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
            example:
              transactionId: "TX-2024-001"
              amount: 250.00
              currency: "USD"
              sourceUserId: 101
              targetUserId: 202
              sourceCurrency: "USD"
              targetCurrency: "EUR"
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/payments/{id}:
    get:
      tags:
        - Payments
      summary: Get payment by ID
      operationId: getPayment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/payments/{id}/history:
    get:
      tags:
        - Payments
      summary: Get state machine transition history for a payment
      operationId: getPaymentHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentHistory'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/payments/{id}/execute:
    post:
      tags:
        - Payments
      summary: Initiate payment processing (triggers INITIATE event)
      operationId: executePayment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: Payment processing started
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/payments/webhook:
    post:
      tags:
        - Payments
      summary: Receive an internal state machine event for a payment
      operationId: handlePaymentEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '200':
          description: Event processed
        '400':
          description: Unknown event or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:

    CreatePaymentRequest:
      type: object
      required:
        - transactionId
        - amount
        - currency
        - sourceUserId
        - targetUserId
        - sourceCurrency
        - targetCurrency
      properties:
        transactionId:
          type: string
          description: Unique idempotency key for this payment
          example: "TX-2024-001"
        amount:
          type: number
          format: decimal
          description: Payment amount in the source currency
          example: 250.00
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code of the payment amount
          example: "USD"
        sourceUserId:
          type: integer
          format: int64
          description: User ID of the payer
          example: 101
        targetUserId:
          type: integer
          format: int64
          description: User ID of the payee
          example: 202
        sourceCurrency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code of the source wallet
          example: "USD"
        targetCurrency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code of the target wallet
          example: "EUR"

    Money:
      type: object
      properties:
        amount:
          type: number
          format: decimal
        currency:
          type: string
          minLength: 3
          maxLength: 3

    Payment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        transactionId:
          type: string
        money:
          $ref: '#/components/schemas/Money'
        state:
          type: string
          description: Current top-level state machine state
          example: "PROCESSING"
        sourceUserId:
          type: integer
          format: int64
        targetUserId:
          type: integer
          format: int64
        sourceCurrency:
          type: string
          minLength: 3
          maxLength: 3
        targetCurrency:
          type: string
          minLength: 3
          maxLength: 3
        processingFee:
          type: number
          format: decimal
          description: Fee charged for this payment (1% of amount)
        netAmount:
          type: number
          format: decimal
          description: Amount after fee deduction
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PaymentHistory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        paymentId:
          type: integer
          format: int64
        fromState:
          type: string
        toState:
          type: string
        event:
          type: string
        timestamp:
          type: string
          format: date-time

    WebhookPayload:
      type: object
      required:
        - paymentId
        - event
      properties:
        paymentId:
          type: integer
          format: int64
        event:
          type: string
          description: PaymentEvent enum value to dispatch
          example: "AUTHORIZE"

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time